+++ controller/Changelog.md
@@ -1,3 +1,9 @@
+# [UNRELEASED]
+
+## Removed
+
+- controller: Remove unused label printing functionality
+
 # [2024.7.0] - 2024-08-02
 
 # [2024.7.0-VALIDATION] - 2024-07-02